requirements.txt
requests
pandas
openpyxl
pdfplumber
spacy
pydrive
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai
pydrive2
Flask
python-dotenv

